refactor: Car 클래스를 별도의 파일로 분리
하나의 파일에는 하나의 클래스만 존재하도록 Car 클래스를 분리함
