+++ src/Car.js
@@ -1,0 +1,17 @@
+import { Random } from '@woowacourse/mission-utils';
+
+class Car {
+  constructor(name) {
+    this.name = name;
+    this.location = 0;
+  }
+
+  move() {
+    const randomNumber = Random.pickNumberInRange(0, 9);
+    if (randomNumber >= 4) {
+      this.location += 1;
+    }
+  }
+}
+
+export default Car;
