+++ src/SettingCars.js
@@ -1,25 +1,12 @@
-import { Console, Random } from '@woowacourse/mission-utils';
+import { Console } from '@woowacourse/mission-utils';
 import Validation from './Validation.js';
 import ErrorHandler from './ErrorHandler.js';
-
-class Car {
-  constructor(name) {
-    this.name = name;
-    this.location = 0;
-  }
-
-  move() {
-    const randomNumber = Random.pickNumberInRange(0, 9);
-    if (randomNumber >= 4) {
-      this.location += 1;
-    }
-  }
-}
+import Car from './Car.js';
 
 class SettingCars {
   async registerCars() {
     const carNameList = await Console.readLineAsync(
-      '경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)'
+      '경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)',
     );
     const carNameArray = carNameList.split(',');
